Relocate all input related functions to Input
