+++ lib/TTT/state.ex
@@ -29,28 +29,6 @@ defmodule State do
     end)
   end
 
-  def user_move_to_internal_state(move) do
-    try do
-      Map.get(
-        %{
-          1 => {0, 0},
-          2 => {0, 1},
-          3 => {0, 2},
-          4 => {1, 0},
-          5 => {1, 1},
-          6 => {1, 2},
-          7 => {2, 0},
-          8 => {2, 1},
-          9 => {2, 2}
-        },
-        move,
-        {:error, :nonexistant_space}
-      )
-    rescue
-      ArgumentError -> {:error, :nonexistant_space}
-    end
-  end
-
   def set_move(move, state) do
     %State{
       player: switch_player(state.player),
