+++ test/input_test.exs
@@ -25,6 +25,28 @@ defmodule InputTest do
     test "returns NAN error message" do
       assert Input.sanitized_move("a") == {:error, :not_a_number}
     end
+
+    test "returns nonexistant space error if input is nan" do
+      assert Input.user_move_to_internal_state("a") ==
+               {:error, :nonexistant_space}
+    end
+
+    test "identifies input as not matching an existing space" do
+      assert Input.user_move_to_internal_state("10") ==
+               {:error, :nonexistant_space}
+    end
+
+    test "identifies space that has been taken by \"X\"", context do
+      assert Input.open?("1", context[:intermediary_board_state]) == {:error, :space_taken}
+    end
+
+    test "identifies space that has been taken by \"O\"", context do
+      assert Input.open?("2", context[:intermediary_board_state]) == {:error, :space_taken}
+    end
+
+    test "identifies if input is not an existing space on the board" do
+      assert Input.space_on_board?("10") == {:error, :nonexistant_space}
+    end
   end
 
   describe "when input is valid" do
@@ -49,6 +71,18 @@ defmodule InputTest do
 
     test "move returns integer version of input" do
       assert Input.sanitized_move("1") == 1
+    end
+
+    test "can convert move to internal state" do
+      assert Input.user_move_to_internal_state(1) == {0, 0}
+    end
+
+    test "identifies input as a space that exists on the board", context do
+      assert Input.space_on_board?(1) == {:ok, :space_on_board}
+    end
+
+    test "confirms space is open", context do
+      assert Input.open?(1, context[:empty_board_state])
     end
 
     test "retrieve gets user's input" do
