+++ test/state_test.exs
@@ -56,20 +56,6 @@ defmodule StateTest do
            }
   end
 
-  test "can convert move to internal state" do
-    assert State.user_move_to_internal_state(1) == {0, 0}
-  end
-
-  test "returns nonexistant space error if input is nan" do
-    assert State.user_move_to_internal_state("a") ==
-             {:error, :nonexistant_space}
-  end
-
-  test "identifies input as not matching an existing space" do
-    assert State.user_move_to_internal_state("10") ==
-             {:error, :nonexistant_space}
-  end
-
   test "switch player symbol from \"X\" to \"O\"", context do
     assert State.switch_player(context[:empty_board_state].player) == "O"
   end
